# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nomi Core",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transactions/{transaction_id}/reverse": {
      "post": {
        "summary": "Reverse",
        "operationId": "reverse_api_v1_transactions__transaction_id__reverse_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReversalInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commitments/{commitment_id}/transactions": {
      "post": {
        "summary": "Pay",
        "operationId": "pay_api_v1_commitments__commitment_id__transactions_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "History",
        "operationId": "history_api_v1_commitments__commitment_id__transactions_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TransactionOut_"
                }
              }
            }
          },
          "400": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "408": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "contact_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contact Id"
          },
          "direction": {
            "type": "string",
            "enum": [
              "receivable",
              "payable"
            ],
            "title": "Direction"
          },
          "original_amount_minor": {
            "type": "integer",
            "title": "Original Amount Minor"
          },
          "balance_minor": {
            "type": "integer",
            "title": "Balance Minor"
          },
          "currency_code": {
            "type": "string",
            "title": "Currency Code"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "lifecycle_status": {
            "type": "string",
            "enum": [
              "open",
              "paid",
              "cancelled"
            ],
            "title": "Lifecycle Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "payment_state": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid"
            ],
            "title": "Payment State"
          },
          "timing_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "no_due_date",
                  "upcoming",
                  "due_soon",
                  "overdue"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Timing State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contact_id",
          "direction",
          "original_amount_minor",
          "balance_minor",
          "currency_code",
          "concept",
          "due_date",
          "lifecycle_status",
          "version",
          "payment_state",
          "timing_state"
        ],
        "title": "CommitmentOut"
      },
...
      "Page_TransactionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[TransactionOut]"
      },
      "PaymentInput": {
        "properties": {
          "type": {
            "type": "string",
            "const": "payment",
            "title": "Type",
            "default": "payment"
          },
          "amount_minor": {
            "type": "integer",
            "maximum": 9000000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 2147483646.0,
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_minor",
          "expected_version",
          "occurred_at"
        ],
        "title": "PaymentInput"
      },
      "PaymentOut": {
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TransactionOut"
          },
          "commitment": {
            "$ref": "#/components/schemas/CommitmentOut"
          }
        },
        "type": "object",
        "required": [
          "transaction",
          "commitment"
        ],
        "title": "PaymentOut"
      },
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "Problem"
      },
...
      "ReversalInput": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "maximum": 2147483646.0,
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "ReversalInput"
      },
...
      "TransactionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "payment",
              "reversal"
            ],
            "title": "Type"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          },
          "currency_code": {
            "type": "string",
            "title": "Currency Code"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reversal_of_transaction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reversal Of Transaction Id"
          },
          "reversed_by_transaction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reversed By Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "amount_minor",
          "currency_code",
          "occurred_at",
          "created_at",
          "note"
        ],
        "title": "TransactionOut"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nomi_session"
      }
    }
  }
}